Refactor: Rename Flagwar's listeners to match
TownyWarXListener sounds more generified than what these listeners actually do. Renamed to match FlagWar and FlagWarConfig.

(Any formatting changes are due solely to the semi-automated refactoring by IntelliJ IDEA.)
